Fix payment date format (#99)
* UpdateFile PaymentsEntryPoint.php (#91)

On line 72, if $paymentDate return DateTime with specific payment date, and it's been format to "Y-m-d\TH:i:sP" the http request will result to a bed http request since the require date format is 'Y-m-d' (without TH:i:sP). The quick fix is to change the format on line 72 to 'Y-,-d' instead of `DateTime::RFC3339`. The better way to make this work is to change the required date format to "Y-m-d\TH:i:sP" so we can continue use `DateTime::RFC3339`.

* Update payments tests for change in payment date format

* Fix typo in payments test file

Co-authored-by: mo6709 <mo6709@gmail.com>